feat: Updated tooltip re-rendering
