+++ package.json
@@ -30,7 +30,6 @@
         "@radix-ui/react-label": "^0.0.14",
         "@radix-ui/react-tooltip": "^0.0.19",
         "@reach/portal": "^0.10.5",
-        "@reduxjs/toolkit": "^1.8.1",
         "@rematch/core": "^2.2.0",
         "@sentry/react": "^6.2.5",
         "@stitches/react": "^0.1.9",
@@ -74,8 +73,11 @@
         "react-modal": "^3.11.2",
         "react-redux": "^7.2.3",
         "react-svg-loader": "^3.0.3",
+        "react-uid": "^2.3.1",
         "react-window": "^1.8.6",
+        "redux": "^4.0.0",
         "regenerator-runtime": "^0.13.7",
+        "reselect": "^4.1.5",
         "set-value": "^3.0.2",
         "storyblok-js-client": "^3.3.1",
         "tailwindcss": "^3.0.0",
@@ -126,7 +128,7 @@
         "ts-node": "^9.1.1",
         "tslint": "^5.18.0",
         "tslint-react": "^4.0.0",
-        "typescript": "^3.7.4",
+        "typescript": "^4.6.3",
         "url-loader": "^2.1.0",
         "webpack": "^4.41.4",
         "webpack-cli": "^3.3.6"
