+++ src/app/components/MoreButton/MoreButtonProperty.tsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import { useSelector } from 'react-redux';
+import { CheckIcon } from '@radix-ui/react-icons';
+import {
+  ContextMenuCheckboxItem,
+  ContextMenuItemIndicator,
+} from '../ContextMenu';
+import { PropertyObject } from '@/types/properties';
+import { mainNodeSelectionValuesSelector } from '@/selectors';
+
+type Props = {
+  value: string;
+  property: PropertyObject;
+  onClick: (properties: PropertyObject | PropertyObject[], isActive: boolean) => void;
+};
+
+export const MoreButtonProperty: React.FC<Props> = ({ value, property, onClick }) => {
+  const mainNodeSelectionValues = useSelector(mainNodeSelectionValuesSelector);
+  const isActive = React.useMemo(() => (
+    mainNodeSelectionValues[property.name] === value
+  ), [value, property, mainNodeSelectionValues]);
+  const handleClick = React.useCallback(() => {
+    onClick(property, isActive);
+  }, [property, isActive, onClick]);
+
+  return (
+    <ContextMenuCheckboxItem
+      key={property.label}
+      checked={isActive}
+      onCheckedChange={handleClick}
+    >
+      <ContextMenuItemIndicator>
+        <CheckIcon />
+      </ContextMenuItemIndicator>
+      {property.label}
+    </ContextMenuCheckboxItem>
+  );
+};
