+++ src/app/components/MoreButton/index.ts
@@ -1,0 +1,1 @@
+export * from './MoreButton';
