+++ src/app/components/TokenTooltip/index.ts
@@ -1,0 +1,1 @@
+export * from './TokenTooltip';
