+++ src/selectors/activeTokenSetSelector.ts
@@ -1,7 +1,7 @@
-import { createSelector } from '@reduxjs/toolkit';
+import { createSelector } from 'reselect';
 import { tokenStateSelector } from './tokenStateSelector';
 
 export const activeTokenSetSelector = createSelector(
   tokenStateSelector,
-  (state) => state.activeTokenSet,
+  (tokenState) => tokenState.activeTokenSet,
 );
