+++ src/selectors/editProhibitedSelector.ts
@@ -1,4 +1,4 @@
-import { createSelector } from '@reduxjs/toolkit';
+import { createSelector } from 'reselect';
 import { tokenStateSelector } from './tokenStateSelector';
 
 export const editProhibitedSelector = createSelector(tokenStateSelector, (state) => state.editProhibited);
