+++ src/selectors/featureFlagsSelector.ts
@@ -1,4 +1,4 @@
-import { createSelector } from '@reduxjs/toolkit';
+import { createSelector } from 'reselect';
 import { uiStateSelector } from './uiStateSelector';
 
 export const featureFlagsSelector = createSelector(
