+++ src/selectors/ignoreFirstPartForStylesSelector.ts
@@ -1,4 +1,4 @@
-import { createSelector } from '@reduxjs/toolkit';
+import { createSelector } from 'reselect';
 import { settingsStateSelector } from './settingsStateSelector';
 
 export const ignoreFirstPartForStylesSelector = createSelector(
