+++ src/selectors/importedTokensSelector.ts
@@ -1,5 +1,5 @@
 import isEqual from 'lodash.isequal';
-import { createSelector } from '@reduxjs/toolkit';
+import { createSelector } from 'reselect';
 import { tokenStateSelector } from './tokenStateSelector';
 
 export const importedTokensSelector = createSelector(
