+++ src/selectors/index.ts
@@ -31,3 +31,4 @@ export * from './showEmptyGroupsSelector';
 export * from './updateModeSelector';
 export * from './displayTypeSelector';
 export * from './collapsedSelector';
+export * from './uiDisabledSelector';
