+++ src/selectors/projectURLSelector.ts
@@ -1,4 +1,4 @@
-import { createSelector } from '@reduxjs/toolkit';
+import { createSelector } from 'reselect';
 import { uiStateSelector } from './uiStateSelector';
 
 export const projectURLSelector = createSelector(
