+++ src/selectors/usedTokenSetSelector.ts
@@ -1,4 +1,4 @@
-import { createSelector } from '@reduxjs/toolkit';
+import { createSelector } from 'reselect';
 import { tokenStateSelector } from './tokenStateSelector';
 
 export const usedTokenSetSelector = createSelector(
